Move warnings_checker and regex checker to separate file
