+++ src/mlx/warnings.py
@@ -9,10 +9,10 @@ import pkg_resources
 import re
 import subprocess
 import sys
-import abc
 from junitparser import JUnitXml, Failure, Error
 import glob
 from setuptools_scm import get_version
+from mlx.warnings_checker import WarningsChecker, RegexChecker
 
 DOXYGEN_WARNING_REGEX = r"(?:((?:[/.]|[A-Za-z]).+?):(-?\d+):\s*([Ww]arning|[Ee]rror)|<.+>:-?\d+(?::\s*([Ww]arning|[Ee]rror))?): (.+(?:(?!\s*(?:[Nn]otice|[Ww]arning|[Ee]rror): )[^/<\n][^:\n][^/\n].+)*)|\s*([Nn]otice|[Ww]arning|[Ee]rror): (.+)\n?"
 doxy_pattern = re.compile(DOXYGEN_WARNING_REGEX)
@@ -21,134 +21,6 @@ SPHINX_WARNING_REGEX = r"(.+?:(?:\d+|None)?):?\s*(DEBUG|INFO|WARNING|ERROR|SEVER
 sphinx_pattern = re.compile(SPHINX_WARNING_REGEX)
 
 __version__ = get_version()
-
-
-class WarningsChecker(object):
-    name = 'checker'
-
-    def __init__(self, verbose=False):
-        ''' Constructor
-
-        Args:
-            name (str): Name of the checker
-            verbose (bool): Enable/disable verbose logging
-        '''
-        self.verbose = verbose
-        self.reset()
-
-    def reset(self):
-        ''' Reset function (resets min, max and counter values) '''
-        self.count = 0
-        self.warn_min = 0
-        self.warn_max = 0
-
-    @abc.abstractmethod
-    def check(self, content):
-        '''
-        Function for counting the number of warnings in a specific text
-
-        Args:
-            content (str): The content to parse
-        '''
-        return
-
-    def set_maximum(self, maximum):
-        ''' Setter function for the maximum amount of warnings
-
-        Args:
-            maximum (int): maximum amount of warnings allowed
-
-        Raises:
-            ValueError: Invalid argument (min limit higher than max limit)
-        '''
-        if self.warn_min == 0:
-            self.warn_max = maximum
-        elif self.warn_min > maximum:
-            raise ValueError("Invalid argument: mininum limit ({min}) is higher than maximum limit ({max}). Cannot enter {value}". format(min=self.warn_min, max=self.warn_max, value=maximum))
-        else:
-            self.warn_max = maximum
-
-    def get_maximum(self):
-        ''' Getter function for the maximum amount of warnings
-
-        Returns:
-            int: Maximum amount of warnings
-        '''
-        return self.warn_max
-
-    def set_minimum(self, minimum):
-        ''' Setter function for the minimum amount of warnings
-
-        Args:
-            minimum (int): minimum amount of warnings allowed
-
-        Raises:
-            ValueError: Invalid argument (min limit higher than max limit)
-        '''
-        if minimum > self.warn_max:
-            raise ValueError("Invalid argument: mininum limit ({min}) is higher than maximum limit ({max}). Cannot enter {value}". format(min=self.warn_min, max=self.warn_max, value=minimum))
-        else:
-            self.warn_min = minimum
-
-    def get_minimum(self):
-        ''' Getter function for the minimum amount of warnings
-
-        Returns:
-            int: Minimum amount of warnings
-        '''
-        return self.warn_min
-
-    def return_count(self):
-        ''' Getter function for the amount of warnings found
-
-        Returns:
-            int: Number of warnings found
-        '''
-        print("{count} {name} warnings found".format(count=self.count, name=self.name))
-        return self.count
-
-    def return_check_limits(self):
-        ''' Function for checking whether the warning count is within the configured limits
-
-        Returns:
-            int: 0 if the amount of warnings is within limits. the count of warnings otherwise
-        '''
-        if self.count > self.warn_max:
-            print("Number of warnings ({count}) is higher than the maximum limit ({max}). Returning error code 1.".format(count=self.count, max=self.warn_max))
-            return self.count
-        elif self.count < self.warn_min:
-            print("Number of warnings ({count}) is lower than the minimum limit ({min}). Returning error code 1.".format(count=self.count, min=self.warn_min))
-            return self.count
-        else:
-            print("Number of warnings ({count}) is between limits {min} and {max}. Well done.".format(count=self.count, min=self.warn_min, max=self.warn_max))
-            return 0
-
-
-class RegexChecker(WarningsChecker):
-    name = 'regex'
-    pattern = None
-
-    def __init__(self, verbose=False):
-        ''' Constructor
-
-        Args:
-            name (str): Name of the checker
-            pattern (str): Regular expression used by the checker in order to find warnings
-        '''
-        super(RegexChecker, self).__init__(verbose=verbose)
-
-    def check(self, content):
-        '''
-        Function for counting the number of warnings in a specific text
-
-        Args:
-            content (str): The content to parse
-        '''
-        matches = re.finditer(self.pattern, content)
-        for match in matches:
-            self.count += 1
-            if self.verbose:
-                print(match.group(0).strip())
 
 
 class SphinxChecker(RegexChecker):
